feat: skills mutations outside
