+++ src/app/[locale]/(authenticated)/skills/components/Form.tsx
@@ -1,12 +1,11 @@
 'use client'
 
-import { Button, ButtonProps, FloatingForm, useToast } from '@/components'
+import { Button, ButtonProps, FloatingForm } from '@/components'
 import SkillsCommonForm from '@/forms/SkillsCommonForm'
-import { CreateSkillDto, useSkills } from '@/sdk'
+import { useSkillsCreate } from '@/hooks'
+import { CreateSkillDto } from '@/sdk'
 
 import { Plus } from '@phosphor-icons/react'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { AxiosError } from 'axios'
 import { useTranslations } from 'next-intl'
 import { forwardRef, useState } from 'react'
 
@@ -15,30 +14,8 @@ export interface SkillsFormProps extends ButtonProps {}
 const SkillsForm = forwardRef<HTMLButtonElement, SkillsFormProps>(
   ({ ...rest }, ref) => {
     const tSkills = useTranslations('skills')
-    const skills = useSkills()
-    const queryClient = useQueryClient()
-    const { toast } = useToast()
     const [isFormOpen, setIsFormOpen] = useState(false)
-
-    const createSkill = useMutation({
-      mutationFn: skills.skillsControllerCreate.bind(skills),
-      mutationKey: ['skillsControllerCreate'],
-      onSuccess: () => {
-        queryClient.invalidateQueries({ queryKey: ['skills'] })
-        setIsFormOpen(false)
-        toast({
-          title: 'Success',
-          description: 'Created successfully',
-        })
-      },
-      onError: ({ response }: AxiosError<any>) => {
-        toast({
-          title: response?.data.name,
-          description: response?.data.message,
-          variant: 'destructive',
-        })
-      },
-    })
+    const createSkill = useSkillsCreate()
 
     return (
       <>
@@ -59,7 +36,11 @@ const SkillsForm = forwardRef<HTMLButtonElement, SkillsFormProps>(
         >
           <SkillsCommonForm
             isLoading={createSkill.isPending}
-            handleSubmit={(data) => createSkill.mutate(data as CreateSkillDto)}
+            handleSubmit={(data) =>
+              createSkill.mutate(data as CreateSkillDto, {
+                onSuccess: () => setIsFormOpen(false),
+              })
+            }
           />
         </FloatingForm>
       </>
