+++ src/hooks/index.ts
@@ -1,5 +1,13 @@
+// queries
 export { useCoursesFindAll } from './queries/course'
 export { useExperiencesFindAll } from './queries/experiences'
 export { useFormationsFindAll } from './queries/formations'
 export { useProjectsFindAll } from './queries/projects'
 export { useSkillsFindAll } from './queries/skills'
+
+// mutations
+export {
+  useSkillsCreate,
+  useSkillsRemove,
+  useSkillsUpdate,
+} from './mutations/skills'
